# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DIEM Agent Workers API",
    "version": "0.2.0",
    "description": "Machine-first x402 micro-workers powered by private Venice models. Not affiliated with or endorsed by Venice.ai."
  },
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
  "paths": {
...
    "/v1/jobs/transcribe-audio": {
      "post": {
        "operationId": "transcribe_audio",
        "summary": "Transcribe a base64-encoded PCM WAV clip of at most 60 seconds into text. Input duration is verified before payment.",
        "description": "Input and provider readiness are checked before x402 payment is requested.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "audioBase64": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 3800000,
                    "pattern": "^[A-Za-z0-9+/]+={0,2}$",
                    "description": "Canonical base64 containing a PCM WAV file of at most 60 seconds."
                  },
                  "language": {
                    "type": "string",
                    "pattern": "^[a-z]{2}$"
                  }
                },
                "required": [
                  "audioBase64"
                ],
                "additionalProperties": false
              },
              "example": {
                "audioBase64": "UklGRkQDAABXQVZFZm10IBAAAAABAAEAQB8AAEAfAAABAAgAZGF0YSADAACAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgA==",
                "language": "en"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed worker result",
            "content": {
              "application/json": {
                "example": {
                  "worker": "transcribe_audio",
                  "result": {
                    "text": "Your report is ready.",
                    "durationSeconds": 1.4
                  },
                  "validation": {
                    "valid": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; no payment requested"
          },
          "402": {
            "description": "x402 payment required"
          },
          "413": {
            "description": "Payload too large; no payment requested"
          },
          "503": {
            "description": "Provider capacity or required model unavailable; no payment requested"
          }
        }
      }
    }
  }
}